chore: 構造体をmodelで定義
